Moved some ordering of the initialization of the shock quantities so that it didn't conflict with Grid_CosmologySim... with CosmologySim...
Put a qualifier for the flux correction to not touch the shock fields,
which aren't conserved quanities.  Same thing in the
ProjectSolutionToParentGrid so that Mach etc. aren't projected up to
parents.

Took out JBPerf stuff from the ShocksHandler.

Removed reference to an old routine in Make.config.objects.

Fixed an incomplete merge in the Make.mach.darwin