Consider "van Leer" slope
(qplus * qminus) / (qplus + qminus)

when monotonizing the reconstruction.  This idea is taken from
ATHENA.  In the Toro hydro tests, it removes most of the oscillatory
behavior but increases the norm to the exact solution in non-AMR
tests.  Removes most of the noise in the Noh 2D test.

Also added a check to keep the left and right Eulerian states to be
contained within the neighboring cell-centered values.  Also taken
from ATHENA.